pyproject.toml
[project]
name = "fastapi-prisma-realworld"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datamodel-code-generator>=0.26.3",
    "email-validator>=2.2.0",
    "fastapi[all,standard]>=0.115.6",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "prisma>=0.15.0",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.1",
    "python-slugify>=8.0.4",
    "ruff>=0.8.4",
    "tzdata>=2024.2",
]


[tool.ruff]
# 1行の最大文字数
line-length = 88

[tool.ruff.lint]
# チェックするエラーの種類
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
# 除外するエラーの種類
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904", # raise ... from ... になっていない場合のエラーを無視
]

# ファイルごとのエラー除外
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

# isort の設定
[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]
